# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateSecretRequest:
    properties:
      login:
        type: string
      metadata:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
...
  api.SecretItemResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      login:
        type: string
      metadata:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  api.SecretResponse:
    properties:
      checksum:
        type: string
      created_at:
        type: string
      login:
        type: string
      metadata:
        type: string
      name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
...
  api.UpdateSecretRequest:
    properties:
      login:
        type: string
      metadata:
        type: string
      password:
        type: string
    type: object
...
info:
  contact: {}
  description: Менеджер паролей GophKeeper
  title: GophKeeper API
  version: "1.0"
paths:
...
  /api/secret:
    get:
      description: Возвращает секреты текущего пользователя без паролей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.SecretItemResponse'
            type: array
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Список секретов пользователя
      tags:
      - secret
    post:
      consumes:
      - application/json
      description: Создаёт секрет текущего пользователя. Имя секрета уникально в рамках
        пользователя
      parameters:
      - description: Данные секрета
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SecretResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "409":
          description: Секрет с таким именем уже существует
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Создание секрета
      tags:
      - secret
  /api/secret/{secret_name}:
    delete:
      description: Удаляет секрет текущего пользователя по имени
      parameters:
      - description: Имя секрета
        in: path
        name: secret_name
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Секрет не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаление секрета
      tags:
      - secret
    get:
      description: Возвращает секрет текущего пользователя по имени
      parameters:
      - description: Имя секрета
        in: path
        name: secret_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SecretResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Секрет не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение секрета
      tags:
      - secret
    put:
      consumes:
      - application/json
      description: Полностью перезаписывает содержимое секрета текущего пользователя,
        найденного по имени
      parameters:
      - description: Имя секрета
        in: path
        name: secret_name
        required: true
        type: string
      - description: Новые данные секрета
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SecretResponse'
        "400":
          description: Неверный формат запроса
          schema:
            type: string
        "401":
          description: Пользователь не аутентифицирован
          schema:
            type: string
        "404":
          description: Секрет не найден
          schema:
            type: string
        "500":
          description: Внутренняя ошибка
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Обновление секрета
      tags:
      - secret
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"